requirements.txt
# --- API & Web Server ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# --- Database & ORM ---
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# --- Configuration ---
pydantic==2.7.1
pydantic-settings==2.2.1

# --- Caching ---
redis==5.0.4

# --- IA & LangGraph ---
langchain==0.2.1
langgraph==0.0.63
langchain-community==0.2.1
langchain-openai  # Pour l'interface, même si on utilise Ollama

# --- NLP ---
spacy==3.7.4
# Le modèle fr_core_news_lg sera installé dans le Dockerfile

# --- Sécurité ---
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# --- Frontend ---
streamlit==1.35.0
plotly==5.22.0
requests==2.31.0

# --- Traitement de fichiers ---
python-docx==1.1.2
PyMuPDF==1.24.5

# --- Tests ---
pytest==8.2.1
httpx==0.27.0
aiosqlite>=0.19.0
faker>=19.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
